nua_orchestrator_remote_v2/pyproject.toml
[tool.poetry]
name = "nua_orchestrator"
version = "0.1.5-1657609595"
description = "Nua orchestrator is a daemon managing Nua commands and services."
authors = ["Jerome Dumonteil <jd@abilian.com>"]
license = "AGPL"
readme = "README.md"

[tool.poetry.scripts]
nua-orchestrator = "nua_orchestrator.main:app"
nua-adm-print-config = "nua_orchestrator.scripts.admin_local:print_local_nua_config"
nua-adm-gen-host-key = "nua_orchestrator.scripts.admin_local:set_new_host_key"
nua-adm-load-host-key = "nua_orchestrator.scripts.admin_local:set_new_host_key_from_file"

[tool.poetry.dependencies]
python = "^3.10"
tomli = "^1.1.0"
wheel = "^0.37.1"
setuptools = "^62.3.2"
tinyrpc = "^1.1.5"
psutil = "^5.9.0"
SQLAlchemy-serializer = "^1.4.1"
SQLAlchemy = "^1.4.36"
py = "^1.11.0"
pyzmq = "^23.1.0"
docker = "^5.0.3"
Fabric = "^2.7.0"
typer = "^0.4.1"
rich = "^12.4.4"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
# Linting
flake8 = "^4.0.1"
flake8-aaa = "0.12.2"
flake8-bandit = "3.0.0"
flake8-bugbear = "^22.4.25"
flake8-cognitive-complexity = "0.1.0"
flake8-comprehensions = "^3.10.0"
flake8-datetimez = "20.10.0"
flake8-functions = "0.0.7"
flake8-mutable = "^1.2.0"
flake8-pytest = "^1.3"
flake8-simplify = "0.19.2"
flake8-super = "0.1.3"
flake8-super-call = "^1.0.0"
flake8-tidy-imports = "^4.8.0"
flake8-tuple = "0.4.1"
bandit = "^1.7.4"
pep8-naming = "^0.13.0"
mastool = "^0.1.3"
# mccabe = "^0.7"

# Formatting
isort = "^5.10.1"
black = "^22.3.0"
docformatter = "^1.4"

# Typechecking
mypy = "^0.961"
types-toml = "^0.10.7"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
